Allow user interaction with the visual effect views while not consuming touches
The UIVisualEffectViews had userInteractionEnabled set to false since they were
preventing touches from getting passed through to the UIScrollView underneath.
However this was also preventing subviews from receiving touches.

To address this I added an extension that overrides `hitTest`. It treats touches
in subviews (technically subviews of the contentView) normally but does not
consider other touches. This allows the superview to process the touch event.
This means that the buttons in the demo are now touchable and you can still
easily scroll around.

Fixes issue #2